cx_example_skill_nodes: rework dummy skiller
Properly name the dummy skiller and also launch it with robot_id param
