+++ cx_example_skill_nodes/src/cx_example_skill_nodes/dummy_skill.cpp
@@ -1,5 +1,5 @@
 /***************************************************************************
- *  DummyMoveSkill.hpp
+ *  DummySkill.cpp
  *
  *  Created: 16 September 2021
  *  Copyright  2021  Ivaylo Doychev
@@ -18,44 +18,46 @@
  *  Read the full text in the LICENSE.GPL file in the doc directory.
  */
 
-#ifndef CX_EXAMPLE_SKILL_NODES__DUMMYMOVESKILL_HPP
-#define CX_EXAMPLE_SKILL_NODES__DUMMYMOVESKILL_HPP
-
-#include <chrono>
-#include <fstream>
-#include <iostream>
-#include <map>
-#include <memory>
-#include <regex>
-#include <string>
-
 #include "rclcpp/rclcpp.hpp"
 #include "rclcpp_action/rclcpp_action.hpp"
 #include "rclcpp_lifecycle/lifecycle_node.hpp"
 
-#include "cx_skill_execution/SkillExecution.hpp"
+#include "cx_example_skill_nodes/dummy_skill.hpp"
 
-#include "cx_msgs/msg/skill_action_exec_info.hpp"
-#include "cx_msgs/msg/skill_execution.hpp"
 #include "lifecycle_msgs/msg/state.hpp"
 
 namespace cx {
+
 using namespace std::chrono_literals;
 
-class DummyMoveSkill : public cx::SkillExecution {
+DummySkill::DummySkill(const std::string &id,
+                               const std::chrono::nanoseconds &rate)
+    : SkillExecution(id, rate) {
+      executor_id_ = "dummy_skiller";
+    }
+using CallbackReturn =
+    rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn;
 
-  using CallbackReturn =
-      rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn;
+CallbackReturn
+DummySkill::on_activate(const rclcpp_lifecycle::State &state) {
+  std::cerr << "DummySkill::on_activate" << std::endl;
+  counter_ = 0;
 
-public:
-  DummyMoveSkill(const std::string &id,
-                 const std::chrono::nanoseconds &rate);
+  return SkillExecution::on_activate(state);
+}
 
-  CallbackReturn on_activate(const rclcpp_lifecycle::State &state);
+void DummySkill::perform_execution() {
+  RCLCPP_INFO_STREAM(get_logger(), "Dummy Execution of ");
+  RCLCPP_INFO_STREAM(get_logger(), action_name_.c_str());
+  for (const auto &param : action_parameters_) {
+    RCLCPP_INFO_STREAM(get_logger(), "\t[" << param << "]");
+  }
 
-  void perform_execution() override;
+  if (counter_++ > 3) {
+    finish_execution(true, 1.0, "completed");
+  } else {
+    send_feedback(counter_ * 0.0, "running");
+  }
+}
 
-  int counter_;
-};
 } // namespace cx
-#endif // !CX_EXAMPLE_SKILL_NODES__DUMMYMOVESKILL_HPP
